Changes to be committed: 	modified:   .amake/cppSizes 	modified:   .vscode/settings.json 	modified:   build/lib/Mlib.a 	modified:   build/obj/Args.o 	modified:   build/obj/Debug.o 	modified:   build/obj/Http.o 	modified:   build/obj/Profile.o 	modified:   build/obj/Sdl2.o 	modified:   build/obj/String.o 	modified:   build/obj/Sys.o 	modified:   build/obj/Threads.o 	modified:   src/cpp/Args.cpp 	modified:   src/cpp/Debug.cpp 	modified:   src/cpp/FileSys.cpp 	modified:   src/cpp/Http.cpp 	modified:   src/cpp/Math.cpp 	modified:   src/cpp/Profile.cpp 	modified:   src/cpp/Sdl2.cpp 	modified:   src/cpp/String.cpp 	modified:   src/cpp/Sys.cpp 	modified:   src/cpp/Threads.cpp 	modified:   src/include/Args.h 	modified:   src/include/Assert.h 	modified:   src/include/Atomic.h 	modified:   src/include/Debug.h 	modified:   src/include/FileSys.h 	modified:   src/include/Math.h 	modified:   src/include/Profile.h 	modified:   src/include/Sdl2.h 	modified:   src/include/Sys.h

+++ src/cpp/String.cpp
@@ -3,11 +3,11 @@ using namespace std;
 
 namespace Mlib::String
 {
-    size_t
-    findN(const string &str, const string &search, size_t n)
+    u64
+    findN(const string &str, const string &search, u64 n)
     {
-        size_t pos = 0;
-        for (size_t i = 0; i < n; ++i)
+        u64 pos = 0;
+        for (u64 i = 0; i < n; ++i)
         {
             pos = str.find(search, pos);
             if (pos == string::npos)
